create commonPage component and refactor home and deals pages
